requirements.txt
asgiref==3.8.1
Django==5.2.1
numpy==2.2.5
pillow==11.2.1
pycryptodome==3.22.0
sqlparse==0.5.3
typing_extensions==4.13.2
gunicorn
